rename linear algebra helper module to more clearly reflect its purpose
