+++ src/lib.rs
@@ -16,7 +16,7 @@ pub mod solvers;
 
 /// private module that contains helper functionality for linear algebra that is not yet
 /// implemented in the nalgebra crate
-mod linear_algebra;
+mod linalg_helpers;
 
 #[cfg(test)]
 mod test_helpers;
